+++ dbus-serialbattery/utils.py
@@ -14,7 +14,7 @@ import serial
 
 
 # CONSTANTS
-DRIVER_VERSION: str = "2.0.20250518dev"
+DRIVER_VERSION: str = "2.0.20250524dev"
 """
 current version of the driver
 """